+++ lib/anagram/anagram_tile.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:anagrammatic/anagram.dart';
+import 'package:anagrammatic/anagram/anagram.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class AnagramTile extends StatefulWidget {