+++ lib/app_flow/home.dart
@@ -1,9 +1,10 @@
-import 'package:anagrammatic/options.dart';
+import 'package:anagrammatic/app_flow/app.dart';
+import 'package:anagrammatic/options/options.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
-import 'package:anagrammatic/backdrop.dart';
-import 'package:anagrammatic/input.dart';
+import 'package:anagrammatic/app_flow/backdrop.dart';
+import 'package:anagrammatic/app_flow/input.dart';
 
 const Duration frontLayerSwitchDuration = Duration(
   milliseconds: 300,
@@ -26,6 +27,12 @@ class _HomeState extends State<Home> {
   bool showBackButton = false;
   bool shouldSystemBackExit = true;
 
+  @override
+  void initState() {
+    super.initState();
+    frontPage = _getInputWidget();
+  }
+
   Input _getInputWidget() {
     return Input(
       transition: _listTransition,
@@ -41,18 +48,13 @@ class _HomeState extends State<Home> {
   }
 
   void _inputTransition() {
-    print('pop');
-    setState(() {
-      showBackButton = false;
-      shouldSystemBackExit = true;
-      frontPage = _getInputWidget();
-    });
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    frontPage = _getInputWidget();
+    if (!AnagrammaticApp.of(context).options.isOpen) {
+      setState(() {
+        showBackButton = false;
+        shouldSystemBackExit = true;
+        frontPage = _getInputWidget();
+      });
+    }
   }
 
   @override