+++ lib/main.dart
@@ -1,4 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:anagrammatic/app.dart';
+import 'package:anagrammatic/app_flow/app.dart';
 
 void main() => runApp(AnagrammaticApp());